+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
+package-lock.json
 .DS_Store
+build/
