+++ index.html
@@ -21,79 +21,15 @@
         <a style="font-size: 12px;" id="showAllDescriptions">Show all descriptions</a>
         <a style="font-size: 12px;" id="hideAllDescriptions">Hide all descriptions</a>
 
-        
+        <% extensions.forEach(ext=> { %>
         <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="BeatsBlox" style="margin-right: 10px" hidden /><a href="https://dev.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/BeatsBlox/index.js%22]#" target="_blank"> BeatsBlox </a></summary>
+            <summary><input type="checkbox" class="extensionOption" name="<%= ext.name %>" style="margin-right: 10px" hidden /><a href="<%= ext.linkUrl %>" target="_blank"> <%= ext.displayName %> </a></summary>
             <p>
-                <input type="hidden" name="BeatsBlox_script" id="BeatsBlox_script" value="https://extensions.netsblox.org/extensions/BeatsBlox/index.js" />
-                BeatsBlox extends Music Functionality within NetsBlox
+                <input type="hidden" name="<%= ext.name %>_script" id="<%= ext.name %>_script" value="<%= ext.scriptUrl %>" />
+                <%= ext.description %>
             </p>
         </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="BetterShare" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/BetterShare/index.js%22]#" target="_blank"> BetterShare </a></summary>
-            <p>
-                <input type="hidden" name="BetterShare_script" id="BetterShare_script" value="https://extensions.netsblox.org/extensions/BetterShare/index.js" />
-                WIP - Provides a few utilities that can make sharing projects easier
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="HandGestures" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/HandGestures/index.js%22]#" target="_blank"> HandGestures </a></summary>
-            <p>
-                <input type="hidden" name="HandGestures_script" id="HandGestures_script" value="https://extensions.netsblox.org/extensions/HandGestures/index.js" />
-                Track 3D hand gestures in images/video using MediaPipe!
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="HideCategories" style="margin-right: 10px" hidden /><a href="https://dev.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/HideCategories/index.js%22]#" target="_blank"> HideCategories </a></summary>
-            <p>
-                <input type="hidden" name="HideCategories_script" id="HideCategories_script" value="https://extensions.netsblox.org/extensions/HideCategories/index.js" />
-                This extension allows you to automatically hide categories and is particularly useful when setting different visible categories for collaborating users.
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="RoboScapeOnline" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/RoboScapeOnline/index.js%22]#" target="_blank"> 🤖 RoboScape Online </a></summary>
-            <p>
-                <input type="hidden" name="RoboScapeOnline_script" id="RoboScapeOnline_script" value="https://extensions.netsblox.org/extensions/RoboScapeOnline/index.js" />
-                Networked robotics simulation in the browser! (WIP)
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="TimeSync" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/TimeSync/index.js%22]#" target="_blank"> TimeSync </a></summary>
-            <p>
-                <input type="hidden" name="TimeSync_script" id="TimeSync_script" value="https://extensions.netsblox.org/extensions/TimeSync/index.js" />
-                calculate time sync info from the NetsBlox server
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="TuneScope" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/TuneScope/index.js%22]#" target="_blank"> TuneScope </a></summary>
-            <p>
-                <input type="hidden" name="TuneScope_script" id="TuneScope_script" value="https://extensions.netsblox.org/extensions/TuneScope/index.js" />
-                Music Notation, Instruments, Drums, Tones, Chords, Tracks, from the University of Virginia (Glen Bull)
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="WebSerial" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/WebSerial/index.js%22]#" target="_blank"> WebSerial </a></summary>
-            <p>
-                <input type="hidden" name="WebSerial_script" id="WebSerial_script" value="https://extensions.netsblox.org/extensions/WebSerial/index.js" />
-                Provides blocks for connecting to a device, e.g. an Arduino, over WebSerial
-            </p>
-        </details>
-        
-        <details class="extension">
-            <summary><input type="checkbox" class="extensionOption" name="WhenKeyPressedLogger" style="margin-right: 10px" hidden /><a href="https://editor.netsblox.org/?extensions=[%22https://extensions.netsblox.org/extensions/WhenKeyPressedLogger/index.js%22]#" target="_blank"> WhenKeyPressedLogger </a></summary>
-            <p>
-                <input type="hidden" name="WhenKeyPressedLogger_script" id="WhenKeyPressedLogger_script" value="https://extensions.netsblox.org/extensions/WhenKeyPressedLogger/index.js" />
-                Logs 'When [key] key pressed' block activations for Ben
-            </p>
-        </details>
-        
+        <% }) %>
 
         <hr />
