+++ index.js
@@ -61,11 +61,10 @@ search.oninput = () => {
         let result = fuzzysort.single(search.value, fullText);
         extension.style.display = (search.value == "" || result)? "block" : "none";
 
-
         if(search.value == "") {
             // Clear highlight when no match
             title.innerHTML = title.innerText;
-            
+
             if(extension.open) {
                 desc.innerHTML = desc.innerText;
             }
